refactor(deploy): Improved multiple application of same Generator
The system we had to handle multiple applications of the same Generator
by detecting and blocking that situation was too simplistic. Sometimes
a Generator needs to know it's being applied multiple times and perform
different actions each time. We now just apply Generators multiple times
and it's the responsibility of the Generator to detect that somehow
(for example by looking at the already existing resources/files) and
act appropriatly.
A special method `filesCopied()` was added to help with that.

Fixes #46