feat(welcome): Welcome App is Capability now and uses new template
The Welcome App was turned into a Capability so we can use the
post-apply feature that Capabilities have.
We're also using the new OS template for the Welcome App that uses
a chained build with nginx using a special configuration for
reverse proxying requests from the frontend to the backend.

Fixes #64 and #47