fix(generators): The welcome-app now properly updates ENV vars
Becuase it would test if the template was already generated and
wouldn't do anything if there was it would actually only work
correctly if the WA was the last capability being added. Otherwise
the ENV for `WELCOME_CONFIG` would keep its intitial value. Now
we always update the `WELCOME_CONFIG` variable.