feat(deploy): Introduced the concept of 'tiers'
In the context of the Creator an "Application" is considered to be
a single code base (a Git repository for example). An application
could have a set of capabilities describing its features. And a set
of capabilities are supposed to only have a single "runtime". This
meant that a single application could only support a single runtime.
This was considered too limiting for the MVP so we needed to come
up with a way to support multiple runtimes in a single application.
The solution was to introduce the concept of "tiers". A tier is a
set of capabilities. A tier can only have a single runtime. But an
application can have multiple tiers. A tier also maps to a sub folder
in the code base, so each tier can have its own set of files.