feat(it): Be more flexible with the error codes we accept
Some frameworks handle certain service errors in a generic
way and make it somewhat more difficult to override them
with the error codes the IT tests are expecting (the ones
defined by the spec for the Boosters). Instead of forcing
the generators to create slightly less idiomatic code we
decided to be more flexible with the error codes we accept
as being valid.